+++ decidim-core/app/helpers/decidim/scopes_helper.rb
@@ -3,6 +3,9 @@
 module Decidim
   # A Helper to render scopes, including a global scope, for forms.
   module ScopesHelper
+    include DecidimFormHelper
+    include TranslatableAttributes
+
     Option = Struct.new(:id, :name)
 
     # Checks if the resource should show its scope or not.
@@ -37,6 +40,21 @@ module Decidim
       root = try(:current_participatory_space)&.scope if root == false
       form.scopes_picker name, options do |scope|
         { url: decidim.scopes_picker_path(root: root, current: scope&.id, field: form.label_for(name)),
+          text: scope_name_for_picker(scope, I18n.t("decidim.scopes.global")) }
+      end
+    end
+
+    # Renders a scopes picker field in a form, not linked to a specific model.
+    # name - name for the input
+    # value - value for the input
+    #
+    # Returns nothing.
+    def scopes_picker_tag(name, value, options = {})
+      root = try(:current_participatory_space)&.scope
+      field = options[:field] || name
+
+      scopes_picker_field_tag name, value, id: options[:id] do |scope|
+        { url: decidim.scopes_picker_path(root: root, current: scope&.id, field: field),
           text: scope_name_for_picker(scope, I18n.t("decidim.scopes.global")) }
       end
     end